Server/tmwine/requirements.txt
absl-py==0.11.0
asgiref==3.4.1
astor==0.8.1
distlib==0.3.4
Django==4.0
djangorestframework==3.13.1
filelock==3.4.2
gast==0.4.0
grpcio==1.35.0
h5py==3.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
mysqlclient==2.1.0
numpy==1.20.0
Pillow==8.4.0
platformdirs==2.4.1
protobuf==3.14.0
pytz==2021.3
six==1.15.0
sqlparse==0.4.2
tensorboard==1.12.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.12.0-py3-none-any.whl
termcolor==1.1.0
virtualenv==20.11.0
Werkzeug==1.0.1
